Extract multi-select dropdown for vacancy filters
